# Rewrite Postgres size query and add `postgresql.relation.{tuples,pages,all_visible}` + toast_size metrics