--- a/postgres/tests/test_relationsmanager.py
+++ b/postgres/tests/test_relationsmanager.py
@@ -3,11 +3,18 @@
 # Licensed under Simplified BSD License (see LICENSE)
 import pytest
 
-from datadog_checks.postgres.relationsmanager import ALL_SCHEMAS, IDX_METRICS, LOCK_METRICS, RelationsManager
+from datadog_checks.postgres.relationsmanager import (
+    ALL_SCHEMAS,
+    IDX_METRICS,
+    LOCK_METRICS,
+    QUERY_PG_CLASS,
+    RelationsManager,
+)
 
 from .common import SCHEMA_NAME
 
 pytestmark = pytest.mark.unit
+default_max_relations = 300
 
 
 @pytest.mark.parametrize(
@@ -47,7 +54,7 @@ pytestmark = pytest.mark.unit
 )
 def test_relations_cases(relations_config, expected_filter):
     query = '{relations}'
-    relations = RelationsManager(relations_config)
+    relations = RelationsManager(relations_config, default_max_relations)
     query_filter = relations.filter_relation_query(query, SCHEMA_NAME)
     assert query_filter == expected_filter
 
@@ -55,7 +62,7 @@ def test_relations_cases(relations_config, expected_filter):
 def test_relation_filter():
     query = "Select foo from bar where {relations}"
     relations_config = [{'relation_name': 'breed', 'schemas': ['public']}]
-    relations = RelationsManager(relations_config)
+    relations = RelationsManager(relations_config, default_max_relations)
 
     query_filter = relations.filter_relation_query(query, SCHEMA_NAME)
     assert (
@@ -67,7 +74,7 @@ def test_relation_filter():
 def test_relation_filter_no_schemas():
     query = "Select foo from bar where {relations}"
     relations_config = [{'relation_name': 'persons', 'schemas': [ALL_SCHEMAS]}]
-    relations = RelationsManager(relations_config)
+    relations = RelationsManager(relations_config, default_max_relations)
 
     query_filter = relations.filter_relation_query(query, SCHEMA_NAME)
     assert query_filter == "Select foo from bar where (( relname = 'persons' ))"
@@ -76,7 +83,7 @@ def test_relation_filter_no_schemas():
 def test_relation_filter_regex():
     query = "Select foo from bar where {relations}"
     relations_config = [{'relation_regex': 'b.*', 'schemas': [ALL_SCHEMAS]}]
-    relations = RelationsManager(relations_config)
+    relations = RelationsManager(relations_config, default_max_relations)
 
     query_filter = relations.filter_relation_query(query, SCHEMA_NAME)
     assert query_filter == "Select foo from bar where (( relname ~ 'b.*' ))"
@@ -85,16 +92,25 @@ def test_relation_filter_regex():
 def test_relation_filter_relkind():
     query = LOCK_METRICS['query'].replace('{metrics_columns}', 'foo')
     relations_config = [{'relation_regex': 'b.*', 'schemas': [ALL_SCHEMAS], 'relkind': ['r', 't']}]
-    relations = RelationsManager(relations_config)
+    relations = RelationsManager(relations_config, default_max_relations)
 
     query_filter = relations.filter_relation_query(query, SCHEMA_NAME)
     assert "AND relkind = ANY(array['r','t'])" in query_filter
 
 
+def test_relation_filter_limit():
+    query = QUERY_PG_CLASS['query']
+    relations_config = [{'relation_regex': '.*', 'schemas': [ALL_SCHEMAS]}]
+    relations = RelationsManager(relations_config, default_max_relations)
+
+    query_filter = relations.filter_relation_query(query, SCHEMA_NAME)
+    assert 'LIMIT 300' in query_filter
+
+
 def test_relkind_does_not_apply_to_index_metrics():
     query = IDX_METRICS['query'].replace('{metrics_columns}', 'foo')
     relations_config = [{'relation_regex': 'b.*', 'schemas': [ALL_SCHEMAS], 'relkind': ['r']}]
-    relations = RelationsManager(relations_config)
+    relations = RelationsManager(relations_config, default_max_relations)
 
     query_filter = relations.filter_relation_query(query, SCHEMA_NAME)
     assert 'relkind' not in query_filter