--- a/postgres/tests/utils.py
+++ b/postgres/tests/utils.py
@@ -12,6 +12,10 @@ requires_over_10 = pytest.mark.skipif(
     POSTGRES_VERSION is None or float(POSTGRES_VERSION) < 10,
     reason='This test is for over 10 only (make sure POSTGRES_VERSION is set)',
 )
+requires_over_11 = pytest.mark.skipif(
+    POSTGRES_VERSION is None or float(POSTGRES_VERSION) < 11,
+    reason='This test is for over 11 only (make sure POSTGRES_VERSION is set)',
+)
 requires_over_14 = pytest.mark.skipif(
     POSTGRES_VERSION is None or float(POSTGRES_VERSION) < 14,
     reason='This test is for over 14 only (make sure POSTGRES_VERSION is set)',