@justiniso 2023-05-04
For my education, the flag `-f "postgres.*idle"`, why is `idle` in there?