@bonnefoa 2023-05-05
> For my education, the flag `-f "postgres.*idle"`, why is `idle` in there?

I wanted to look at the backend process matching the psql session which will run the query I'm testing. 
If you check all processes with a postgres's name, you're gonna match all postgres's processes + psql backend process
```
pgrep -a -f "postgres:"
1017179 postgres: checkpointer
1017180 postgres: background writer
1017182 postgres: walwriter
1017183 postgres: autovacuum launcher
1017184 postgres: logical replication launcher
1826329 postgres: postgres postgres [local] idle
```
An idle session will be tagged as `idle` in the process name so the "postgres.*idle" allowed me to match the backend process of the psql session I have opened. 
```
pgrep -a -f "postgres.*idle"
1826329 postgres: postgres postgres [local] idle
```