@bonnefoa 2023-06-08
Here are perf before and after the patch
![old_perf](https://github.com/DataDog/integrations-core/assets/30333/071d9827-3aa4-4998-bcdb-7491315dd90a)
![new_perf](https://github.com/DataDog/integrations-core/assets/30333/5e8b9c20-cf88-45cb-9dcc-c02464c9d1d9)
Time spent InterpExpr went from 87% to 56%